+++ pallets/rewards/src/mechanism.rs
@@ -1,0 +1,68 @@
+use sp_runtime::ArithmeticError;
+
+pub mod base;
+
+pub trait RewardAccount {
+	type Balance: Copy;
+
+	fn staked(&self) -> Self::Balance;
+}
+
+pub trait RewardGroup {
+	type Balance: Copy;
+
+	fn total_staked(&self) -> Self::Balance;
+}
+
+pub trait RewardMechanism {
+	type Group: RewardGroup<Balance = Self::Balance>;
+	type Account: RewardAccount<Balance = Self::Balance>;
+	type Currency;
+	type Balance;
+
+	fn reward_group(group: &mut Self::Group, amount: Self::Balance) -> Result<(), ArithmeticError>;
+
+	fn deposit_stake(
+		account: &mut Self::Account,
+		currency: &mut Self::Currency,
+		group: &mut Self::Group,
+		amount: Self::Balance,
+	) -> Result<(), ArithmeticError>;
+
+	fn withdraw_stake(
+		account: &mut Self::Account,
+		currency: &mut Self::Currency,
+		group: &mut Self::Group,
+		amount: Self::Balance,
+	) -> Result<(), ArithmeticError>;
+
+	fn compute_reward(
+		account: &mut Self::Account,
+		currency: &Self::Currency,
+		group: &Self::Group,
+	) -> Result<Self::Balance, ArithmeticError>;
+
+	fn claim_reward(
+		account: &mut Self::Account,
+		currency: &Self::Currency,
+		group: &Self::Group,
+	) -> Result<Self::Balance, ArithmeticError>;
+
+	fn move_currency(
+		currency: &mut Self::Currency,
+		prev_group: &mut Self::Group,
+		next_group: &mut Self::Group,
+	) -> Result<(), MoveCurrencyError>;
+}
+
+#[derive(Debug)]
+pub enum MoveCurrencyError {
+	Arithmetic(ArithmeticError),
+	MaxMovements,
+}
+
+impl From<ArithmeticError> for MoveCurrencyError {
+	fn from(e: ArithmeticError) -> MoveCurrencyError {
+		Self::Arithmetic(e)
+	}
+}
