+++ runtime/development/src/lib.rs
@@ -1525,6 +1525,12 @@ impl pallet_rewards::Config for Runtime {
 	type PalletId = RewardsPalletId;
 	type Rate = FixedI128;
 	type RewardCurrency = RewardCurrency;
+	type RewardMechanism = pallet_rewards::mechanism::base::Mechanism<
+		Balance,
+		IBalance,
+		FixedI128,
+		MaxCurrencyMovements,
+	>;
 	type SignedBalance = IBalance;
 }
